### 2022-12-09 3d0f0fd
Prep for animated pattern plots and new example
