+++ examples/farfield_analysis.py
@@ -1,0 +1,322 @@
+#!/usr/bin/env python
+# coding: utf-8
+"""
+Modelling Different Farfield Polarisations for an Aperture
+=============================================================
+
+This example uses the frequency domain :func:`lyceanem.models.frequency_domain.calculate_farfield` function to predict
+the farfield pattern for a linearly polarised aperture. This could represent an antenna array without any beamforming
+weights.
+
+
+"""
+import numpy as np
+import open3d as o3d
+import copy
+
+# %%
+# Setting Farfield Resolution and Wavelength
+# -------------------------------------------
+# LyceanEM uses Elevation and Azimuth to record spherical coordinates, ranging from -180 to 180 degrees in azimuth,
+# and from -90 to 90 degrees in elevation. In order to launch the aperture projection function, the resolution in
+# both azimuth and elevation is requried.
+# In order to ensure a fast example, 37 points have been used here for both, giving a total of 1369 farfield points.
+#
+# The wavelength of interest is also an important variable for antenna array analysis, so we set it now for 10GHz,
+# an X band aperture.
+
+az_res = 37
+elev_res = 37
+wavelength = 3e8 / 10e9
+
+# %%
+# Generating consistent aperture to explore farfield polarisations, and rotating the source
+# ----------------------------------------------------------------------------------------------
+
+from lyceanem.base_classes import points,structures,antenna_structures
+
+from lyceanem.geometry.targets import meshedHorn
+antenna_height=4*wavelength
+antenna_width=4*wavelength
+antenna_diameter=(((antenna_height*0.5)**2+(antenna_width*0.5)**2)**0.5)*2
+structure,array_points=meshedHorn(antenna_height, antenna_width, 4*wavelength, 0.05*wavelength,np.radians(10),wavelength*0.5)
+
+horn_antenna=antenna_structures(structures(solids=[structure]), points(points=[array_points]))
+aperture_points=np.asarray(array_points.points).shape[0]
+
+antenna_farfield_distance=((antenna_diameter)**2)/wavelength
+step_num=100
+sep_scale=np.logspace(0,4,step_num)*wavelength
+from lyceanem.models.frequency_domain import calculate_scattering
+
+receive_point_spread=0.1*wavelength
+receive_aperture_height=wavelength*8
+receive_aperture_width=wavelength*8
+receive_diameter=(((receive_aperture_height*0.5)**2+(receive_aperture_width*0.5)**2)**0.5)*2
+receive_farfield_distance=((receive_diameter)**2)/wavelength
+width_point_num=np.ceil(receive_aperture_width/(receive_point_spread)).astype(int)
+height_point_num=np.ceil(receive_aperture_height/(receive_point_spread)).astype(int)
+
+Poynting_vector_distance0p1=np.zeros((width_point_num,height_point_num,step_num))
+sample_angle_distance0p1=np.zeros((width_point_num,height_point_num,step_num))
+for sep_index in range(len(sep_scale)):
+    seperation_distance=sep_scale[sep_index]
+    sink_points=o3d.geometry.PointCloud()
+    sink_cords=np.zeros((width_point_num*height_point_num,3))
+    sink_normals=np.zeros((width_point_num*height_point_num,3))
[…]
