### 2022-12-15 57359a9
Added Initial Satellite Emissions Modelling using Folium and Skyfield
