+++ docs/examples/08_beamforming_architecture.py
@@ -26,8 +26,8 @@ import copy
 # The wavelength of interest is also an important variable for antenna array analysis, so we set it now for 10GHz,
 # an X band aperture.
 
-az_res = 361
-elev_res = 181
+az_res = 181
+elev_res = 37
 wavelength = 3e8 / 10e9
 
 # %%
