+++ docs/examples/09_satellite_channel_modelling.py
@@ -1,0 +1,377 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Satellite Channel Modelling
+======================================================
+
+This example uses the frequency domain :func:`lyceanem.models.frequency_domain.calculate_scattering` function to
+predict the scattering parameters for the frequency and environment included in the model. This allows the effects of antenna polarisation,
+
+
+"""
+
+# %%
+# Define Map and Ground Station Location and list of Satellites using skyfield API to load starlink satellites.
+# ~~~~~~~
+
+import folium
+
+interest_center = [51.458729, -2.602153]  # location of Bristol Physics Observatory
+
+m = folium.Map(
+    location=[
+        interest_center[0],
+        interest_center[1],
+    ],
+    zoom_start=6,
+    tiles="Stamen Terrain",
+)
+
+import numpy as np
+import copy
+
+from skyfield.api import load, EarthSatellite, wgs84
+
+BristolPhysicsObservatory = wgs84.latlon(interest_center[0], interest_center[1], 0)
+
+receiver = BristolPhysicsObservatory
+
+# Build the time range `t` over which to plot, plus other values.
+
+ts = load.timescale()
+
+time_points = 2 * 60  # track over 10 minutes
+# change over time
+t = ts.utc(2022, 12, 15, 13, 4, range(0, time_points))
+timebase = t.utc_datetime()
+
+# %%
+# GeoJSON
+# ~~~~~~~
+# The Folium plugin `TimestampedGeoJson`_ will be used to plot our tracks using timestamped
+# GeoJSONs. As a result, we want to convert our data into `GeoJSON format`_. Firstly, we create
+# our feature collection which we will append our features to.
+#
+# .. _GeoJSON format: https://geojson.org/
+# .. _TimestampedGeoJson:  https://python-visualization.github.io/folium/plugins.html
+
+geo_features = list()
+geo_json = {
+    "type": "FeatureCollection",
+    "features": geo_features,
+}
+# %%
+# Plotting Tracks of Satellites visible to Physics Observatory
+# ~~~~~~~~~~~~~~~
+# Now we append our tracks to our feature collection list. We define `colour_iter` which will
+# allow us to cycle through different colours when mapping our tracks.
+
+from collections import defaultdict
+from itertools import cycle
+
+colour_iter = iter(
+    cycle(
+        [
+            "red",
+            "blue",
+            "green",
+            "purple",
+            "orange",
+            "darkred",
+            "#0909FF",
+            "#F70D1A",
+            "#FF6700",
+            "lightgreen",
+            "#0AFFFF",
+            "#12AD2B",
+            "#E2F516",
+            "#FFFF00",
+            "#F52887",
+        ]
+    )
+)
+colour = defaultdict(lambda: next(colour_iter))
+
+trail_Size = 7  # trail_Size is the number of timestamps we want track to trail for
+from io import BytesIO
+import base64
+
+
+# %%
+# Define Starlinks Tracks
+# ~~~~~~~
+# Convinience function to limit the tracks to a set number of the nearest N satellites.
+
+def starlink_tracks(pointofinterest, timebase, num_satellites=5):
+    starlink_url = "https://celestrak.com/NORAD/elements/supplemental/starlink.txt"
+    starlinks = load.tle_file(starlink_url)
+    tracks = []
+    for sat in starlinks:
+        pos = sat.at(timebase)
+        tracks.append(pos)
+        latcheck = np.any(np.isnan(wgs84.subpoint_of(pos).latitude.degrees))
+        loncheck = np.any(np.isnan(wgs84.subpoint_of(pos).longitude.degrees))
+        if latcheck | loncheck:
[…]
