### 2022-12-19 25c332a
Added Initial Satellite Emissions Modelling using Folium and Skyfield
