+++ docs/source/auto_examples/05_array_beamforming.rst
@@ -2,7 +2,7 @@
 .. DO NOT EDIT.
 .. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
 .. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
-.. "auto_examples/05_array_beamforming.py"
+.. "auto_examples\05_array_beamforming.py"
 .. LINE NUMBERS ARE GIVEN BELOW.
 
 .. only:: html
@@ -195,19 +195,11 @@ and :func:`lyceanem.electromagnetics.beamforming.WavefrontWeights`
 
 .. rst-class:: sphx-glr-script-out
 
- Out:
-
  .. code-block:: none
 
-    /home/timtitan/Documents/10-19-Research-Projects/14-Electromagnetics-Modelling/14.04-Python-Development/LyceanEM/lyceanem/electromagnetics/empropagation.py:3604: ComplexWarning: Casting complex values to real discards the imaginary part
-      global_vector[0] = (
-    /home/timtitan/Documents/10-19-Research-Projects/14-Electromagnetics-Modelling/14.04-Python-Development/LyceanEM/lyceanem/electromagnetics/empropagation.py:3609: ComplexWarning: Casting complex values to real discards the imaginary part
-      global_vector[1] = (
-    /home/timtitan/Documents/10-19-Research-Projects/14-Electromagnetics-Modelling/14.04-Python-Development/LyceanEM/lyceanem/electromagnetics/empropagation.py:3614: ComplexWarning: Casting complex values to real discards the imaginary part
-      global_vector[2] = (
-    /home/timtitan/anaconda3/envs/EMDevelopment/lib/python3.8/site-packages/numba/cuda/cudadrv/devicearray.py:885: NumbaPerformanceWarning: Host array used in CUDA kernel will incur copy overhead to/from device.
+    C:\Users\lycea\anaconda3\envs\cusignal-dev\lib\site-packages\numba\cuda\cudadrv\devicearray.py:885: NumbaPerformanceWarning: Host array used in CUDA kernel will incur copy overhead to/from device.
       warn(NumbaPerformanceWarning(msg))
-    /home/timtitan/Documents/10-19-Research-Projects/14-Electromagnetics-Modelling/14.04-Python-Development/LyceanEM/lyceanem/electromagnetics/beamforming.py:1080: RuntimeWarning: divide by zero encountered in log10
+    C:\Users\lycea\PycharmProjects\LyceanEM-Python\lyceanem\electromagnetics\beamforming.py:1167: RuntimeWarning: divide by zero encountered in log10
       logdata = 10 * np.log10(data)
 
 
@@ -247,12 +239,10 @@ and :func:`lyceanem.electromagnetics.beamforming.WavefrontWeights`
 
 .. rst-class:: sphx-glr-script-out
 
- Out:
-
  .. code-block:: none
 
     Steering Effciency of 9.1%
-    /home/timtitan/Documents/10-19-Research-Projects/14-Electromagnetics-Modelling/14.04-Python-Development/LyceanEM/docs/examples/05_array_beamforming.py:122: RuntimeWarning: divide by zero encountered in log10
+    C:\Users\lycea\PycharmProjects\LyceanEM-Python\docs\examples\05_array_beamforming.py:122: RuntimeWarning: divide by zero encountered in log10
       np.max(10 * np.log10(directivity_map[:, :, 2]))
     Maximum Directivity of 23.0 dBi
 
@@ -262,28 +252,23 @@ and :func:`lyceanem.electromagnetics.beamforming.WavefrontWeights`
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** ( 4 minutes  14.648 seconds)
[…]
