+++ docs/source/auto_examples/01_aperture_projection.rst
@@ -2,7 +2,7 @@
 .. DO NOT EDIT.
 .. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
 .. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
-.. "auto_examples/01_aperture_projection.py"
+.. "auto_examples\01_aperture_projection.py"
 .. LINE NUMBERS ARE GIVEN BELOW.
 
 .. only:: html
@@ -207,67 +207,6 @@ also as an open3d point cloud. This allows easy visualisation using :func:`open3
 
 
 
-.. rst-class:: sphx-glr-script-out
-
- Out:
-
- .. code-block:: none
-
-    [Open3D INFO]   -- Mouse view control --
-    [Open3D INFO]     Left button + drag         : Rotate.
-    [Open3D INFO]     Ctrl + left button + drag  : Translate.
-    [Open3D INFO]     Wheel button + drag        : Translate.
-    [Open3D INFO]     Shift + left button + drag : Roll.
-    [Open3D INFO]     Wheel                      : Zoom in/out.
-    [Open3D INFO] 
-    [Open3D INFO]   -- Keyboard view control --
-    [Open3D INFO]     [/]          : Increase/decrease field of view.
-    [Open3D INFO]     R            : Reset view point.
-    [Open3D INFO]     Ctrl/Cmd + C : Copy current view status into the clipboard.
-    [Open3D INFO]     Ctrl/Cmd + V : Paste view status from clipboard.
-    [Open3D INFO] 
-    [Open3D INFO]   -- General control --
-    [Open3D INFO]     Q, Esc       : Exit window.
-    [Open3D INFO]     H            : Print help message.
-    [Open3D INFO]     P, PrtScn    : Take a screen capture.
-    [Open3D INFO]     D            : Take a depth capture.
-    [Open3D INFO]     O            : Take a capture of current rendering settings.
-    [Open3D INFO] 
-    [Open3D INFO]   -- Render mode control --
-    [Open3D INFO]     L            : Turn on/off lighting.
-    [Open3D INFO]     +/-          : Increase/decrease point size.
-    [Open3D INFO]     Ctrl + +/-   : Increase/decrease width of geometry::LineSet.
-    [Open3D INFO]     N            : Turn on/off point cloud normal rendering.
-    [Open3D INFO]     S            : Toggle between mesh flat shading and smooth shading.
-    [Open3D INFO]     W            : Turn on/off mesh wireframe.
-    [Open3D INFO]     B            : Turn on/off back face rendering.
-    [Open3D INFO]     I            : Turn on/off image zoom in interpolation.
-    [Open3D INFO]     T            : Toggle among image render:
-    [Open3D INFO]                    no stretch / keep ratio / freely stretch.
-    [Open3D INFO] 
-    [Open3D INFO]   -- Color control --
-    [Open3D INFO]     0..4,9       : Set point cloud color option.
-    [Open3D INFO]                    0 - Default behavior, render point color.
-    [Open3D INFO]                    1 - Render point color.
-    [Open3D INFO]                    2 - x coordinate as color.
-    [Open3D INFO]                    3 - y coordinate as color.
-    [Open3D INFO]                    4 - z coordinate as color.
-    [Open3D INFO]                    9 - normal as color.
-    [Open3D INFO]     Ctrl + 0..4,9: Set mesh color option.
-    [Open3D INFO]                    0 - Default behavior, render uniform gray color.
-    [Open3D INFO]                    1 - Render point color.
-    [Open3D INFO]                    2 - x coordinate as color.
[…]

[…]
