### 2022-12-19 29d468f
Added Initial Satellite Emissions Modelling using Folium and Skyfield
