### 2022-12-19 0bed4fa
Added Initial Satellite Emissions Modelling using Folium and Skyfield
