+++ docs/source/auto_examples/08_beamforming_architecture.py
@@ -1,0 +1,153 @@
+#!/usr/bin/env python
+# coding: utf-8
+"""
+Array Beamforming
+======================================================
+
+This example uses the frequency domain :func:`lyceanem.models.frequency_domain.calculate_farfield` function to predict
+the farfield patterns for a linearly polarised aperture with multiple elements. This is then beamformed to all farfield points using multiple open loop beamforming algorithms to attemp to 'map' out the acheivable beamforming for the antenna array using :func:`lyceanem.electromagnetics.beamforming.MaximumDirectivityMap`.
+
+The Steering Efficiency can then be evaluated using :func:`lyceanem.electromagnetics.beamforming.Steering_Efficiency` for the resultant achieved beamforming.
+
+
+"""
+import numpy as np
+import open3d as o3d
+import copy
+
+# %%
+# Setting Farfield Resolution and Wavelength
+# -------------------------------------------
+# LyceanEM uses Elevation and Azimuth to record spherical coordinates, ranging from -180 to 180 degrees in azimuth,
+# and from -90 to 90 degrees in elevation. In order to launch the aperture projection function, the resolution in
+# both azimuth and elevation is requried.
+# In order to ensure a fast example, 37 points have been used here for both, giving a total of 1369 farfield points.
+#
+# The wavelength of interest is also an important variable for antenna array analysis, so we set it now for 10GHz,
+# an X band aperture.
+
+az_res = 181
+elev_res = 37
+wavelength = 3e8 / 10e9
+
+# %%
+# Geometries
+# ------------------------
+# In order to make things easy to start, an example geometry has been included within LyceanEM for a UAV, and the
+# :class:`open3d.geometry.TriangleMesh` structures can be accessed by importing the data subpackage
+import lyceanem.tests.reflectordata as data
+
+body, array, source_coords = data.exampleUAV(10e9)
+
+# %%
+# Visualise the Resultant UAV and Array
+# ---------------------------------------
+# :func:`open3d.visualization.draw_geometries` can be used to visualise the open3d data
+# structures :class:`open3d.geometry.PointCloud` and :class:`open3d.geometry.PointCloud`
+
+mesh_frame = o3d.geometry.TriangleMesh.create_coordinate_frame(
+    size=0.5, origin=[0, 0, 0]
+)
+o3d.visualization.draw_geometries([body, array, source_coords, mesh_frame])
+
+# %%
+# .. image:: ../_static/UAVArraywithPoints.png
+
+
+from lyceanem.base_classes import points,structures,antenna_structures
+aperture=points([source_coords])
+blockers = structures([body, array])
+uav_structure=antenna_structures(blockers, aperture)
+#put rotations here
+
+
+surface_array = copy.deepcopy(array)
+surface_array.triangles = o3d.utility.Vector3iVector(
+    np.asarray(array.triangles)[: len(array.triangles) // 2, :]
+)
+surface_array.triangle_normals = o3d.utility.Vector3dVector(
+    np.asarray(array.triangle_normals)[: len(array.triangle_normals) // 2, :]
+)
+array_surface_area=surface_array.get_surface_area()
+maximum_aperture_efficiency=(4*np.pi*array_surface_area)/(wavelength**2)
+# %%
+# Model Farfield Array Patterns
+# -------------------------------
[…]
