+++ docs/source/auto_examples/07_aperture_farfield_polarisation.rst
@@ -1,0 +1,411 @@
+
+.. DO NOT EDIT.
+.. THIS FILE WAS AUTOMATICALLY GENERATED BY SPHINX-GALLERY.
+.. TO MAKE CHANGES, EDIT THE SOURCE PYTHON FILE:
+.. "auto_examples\07_aperture_farfield_polarisation.py"
+.. LINE NUMBERS ARE GIVEN BELOW.
+
+.. only:: html
+
+    .. note::
+        :class: sphx-glr-download-link-note
+
+        Click :ref:`here <sphx_glr_download_auto_examples_07_aperture_farfield_polarisation.py>`
+        to download the full example code
+
+.. rst-class:: sphx-glr-example-title
+
+.. _sphx_glr_auto_examples_07_aperture_farfield_polarisation.py:
+
+
+Modelling Different Farfield Polarisations for an Aperture
+=============================================================
+
+This example uses the frequency domain :func:`lyceanem.models.frequency_domain.calculate_farfield` function to predict
+the farfield pattern for a linearly polarised aperture. This could represent an antenna array without any beamforming
+weights.
+
+.. GENERATED FROM PYTHON SOURCE LINES 13-17
+
+.. code-block:: default
+
+    import numpy as np
+    import open3d as o3d
+    import copy
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 18-27
+
+Setting Farfield Resolution and Wavelength
+-------------------------------------------
+LyceanEM uses Elevation and Azimuth to record spherical coordinates, ranging from -180 to 180 degrees in azimuth,
+and from -90 to 90 degrees in elevation. In order to launch the aperture projection function, the resolution in
+both azimuth and elevation is requried.
+In order to ensure a fast example, 37 points have been used here for both, giving a total of 1369 farfield points.
+
+The wavelength of interest is also an important variable for antenna array analysis, so we set it now for 10GHz,
+an X band aperture.
+
+.. GENERATED FROM PYTHON SOURCE LINES 27-32
+
+.. code-block:: default
+
+
+    az_res = 37
+    elev_res = 37
+    wavelength = 3e8 / 10e9
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 33-35
+
+Generating consistent aperture to explore farfield polarisations, and rotating the source
+----------------------------------------------------------------------------------------------
+
+.. GENERATED FROM PYTHON SOURCE LINES 35-46
+
+.. code-block:: default
+
+
+    from lyceanem.base_classes import points,structures,antenna_structures
+
+    from lyceanem.geometry.targets import meshedHorn
+    structure,array_points=meshedHorn(3*wavelength, 1*wavelength, 4*wavelength, 0.05*wavelength,np.radians(10),wavelength*0.5)
+
+    horn_antenna=antenna_structures(structures(solids=[structure]), points(points=[array_points]))
+
+
+    from lyceanem.models.frequency_domain import calculate_farfield
+
+
+
+
+
+
+
+
+.. GENERATED FROM PYTHON SOURCE LINES 47-48
+
+The first source polarisation is based upon the u-vector of the source point. When the excitation_function method of the antenna structure class is used, it will calculate the appropriate polarisation vectors based upon the local normal vectors.
+
+.. GENERATED FROM PYTHON SOURCE LINES 48-63
+
[…]

[…]
