### 2022-12-19 95096e6
Added Initial Satellite Emissions Modelling using Folium and Skyfield
