### 2022-12-19 bbe986a
Added Initial Satellite Emissions Modelling using Folium and Skyfield
