### 2022-12-19 e669202
Added Initial Satellite Emissions Modelling using Folium and Skyfield
