+++ docs/examples/09_satellite_channel_modelling.py
@@ -372,6 +372,7 @@ TimestampedGeoJson(
 
 # %%
 
+
 m
 
 
