### 2022-12-19 e3cc1b3
Added Initial Satellite Emissions Modelling using Folium and Skyfield
