### 2023-01-18 6750fb9
Updated Antenna Structure Visualisation to use new Open3D method.
