### 2023-01-20 089dddf
Updated Antenna Structure Visualisation to use new Open3D method.
