### 2023-01-29 #4911 @arfon
:wave: @aaronyd @josephhardinee, @IgorBaratta would any of you be willing to review this submission for JOSS? The submission under consideration is LyceanEM, a A Python package for virtual prototyping of antenna arrays, time and frequency domain channel modelling (https://github.com/LyceanEM/LyceanEM-Python)

Many thanks
Arfon
