+++ lyceanem/electromagnetics/empropagation.py
@@ -1140,19 +1140,19 @@ def freqdomainkernal(
                 scatter_index = i
 
         # print(cu_ray_num,source_sink_index[cu_ray_num,0],source_sink_index[cu_ray_num,1])
-        wave_vector = (2.0 * cmath.pi) / wavelength
+        wave_vector = (2.0 * cmath.pi) / wavelength[0]
         # scatter_coefficient=(1/(4*cmath.pi))**(complex(scatter_index))
         if scatter_index == 0:
             loss = cmath.exp(lengths * wave_vector * 1j) * (
-                wavelength / (4 * cmath.pi * lengths)
+                wavelength[0] / (4 * cmath.pi * lengths)
             )
         elif scatter_index == 1:
             loss = cmath.exp(lengths * wave_vector * 1j) * (
-                wavelength / (4 * cmath.pi * lengths)
+                wavelength[0] / (4 * cmath.pi * lengths)
             )
         elif scatter_index == 2:
             loss = cmath.exp(lengths * wave_vector * 1j) * (
-                wavelength / (4 * cmath.pi * lengths)
+                wavelength[0] / (4 * cmath.pi * lengths)
             )
 
         ray_component[0] *= loss
@@ -2134,79 +2134,169 @@ def EMGPUFreqDomain(source_num, sink_num, full_index, point_information, wavelen
         the resultant scattering network for the provided ray paths
 
     """
+    #ctx = cuda.current_context()
+    #ctx.reset()
     free_mem, total_mem = cuda.current_context().get_memory_info()
-    max_mem = np.ceil(free_mem * 0.5).astype(int)
-
+    max_mem = np.ceil(free_mem).astype(np.int64)
     ray_num = full_index.shape[0]
     threads_in_block = 256
-    max_blocks = 65535
-    maximum_chunk_size = 2 ** 8
-    path_lengths = np.zeros((ray_num), dtype=np.float32)
-    scattering_network = np.zeros((source_num, sink_num, 3, 2), dtype=np.float64)
-    d_scattering_network = cuda.device_array(
-        (
-            scattering_network.shape[0],
-            scattering_network.shape[1],
-            scattering_network.shape[2],
-        ),
-        dtype=np.complex64,
-    )
-    d_scattering_network = cuda.to_device(scattering_network)
-    d_point_information = cuda.device_array(
-        point_information.shape[0], dtype=base_types.scattering_t
-    )
-    d_point_information = cuda.to_device(point_information)
-    d_wavelength = cuda.device_array((1), dtype=np.complex64)
-    d_wavelength = cuda.to_device(
-        np.csingle(np.ones((1), dtype=np.complex64) * wavelength)
-    )
     # divide in terms of a block for each source, then
     depthslice, _ = targettingindex(copy.deepcopy(full_index))
-    depthslice[:, 0] -= 1
-    depthslice[:, 1] -= source_num + 1
-    d_target_index = cuda.device_array(
-        (depthslice.shape[0], depthslice.shape[1]), dtype=np.int64
-    )
-    d_target_index = cuda.to_device(depthslice)
-    d_full_index = cuda.device_array(
-        (full_index.shape[0], full_index.shape[1]), dtype=np.int64
-    )
-    # d_paths=cuda.device_array((path_lengths.shape[0]),dtype=np.float32)
-    # d_polar_c=cuda.device_array((polar_coefficients.shape),dtype=np.complex64)
-    # paths=cp.zeros((path_lengths.shape[0]),dtype=np.float32)
-    # polar_c=cp.zeros((polar_coefficients.shape),dtype=np.complex64)
-    d_full_index = cuda.to_device(full_index)
-    # Here, we choose the granularity of the threading on our device. We want
-    # to try to cover the entire workload of rays and targets with simulatenous threads, so we'll
[…]
