### 2023-03-08 #5234 @editorialbot
[REVIEW]: LyceanEM: A python package for virtual prototyping of antenna arrays, time and frequency domain channel modelling
**Submitting author:** @LyceanEM (<a href="http://orcid.org/0000-0002-8085-1264">Timothy Pelham</a>)
**Repository:** https://github.com/LyceanEM/LyceanEM-Python
**Branch with paper.md** (empty if default branch):
**Version:** 0.04
