### 2023-03-22 12b3de5
Bugfix for mesh_conversion and added import and export functions from antenna pattern class to .ffe format for FEKO.
