+++ lyceanem/base_classes.py
@@ -770,6 +770,106 @@ class antenna_pattern(object3d):
             self.pattern[:, :, 0] = Ea.transpose() + norm
             self.pattern[:, :, 1] = Eb.transpose() + norm
 
+    def import_ffe(self,file_path):
+        """
+        tested for .ffe files with theta in the range 0 to 180 degrees, and phi in the range -180 to 180 degrees
+        Parameters
+        ----------
+        file_path
+
+        Returns
+        -------
+
+        """
+        import copy
+
+        # Open the file and read the comments
+        with open(file_path, 'r') as f:
+            comments = f.readlines()
+
+            # Extract lines starting with #
+        comment_lines = [line.strip() for line in comments if line.startswith('#')]
+
+        # Process the comment lines for keywords
+        for line in comment_lines:
+            if 'source' in line.lower():
+                antenna_name = line.split(':')[1].strip()
+            elif 'result type' in line.lower():
+                pattern_type = line.split(':')[1].strip()
+            elif "frequency" in line.lower():
+                frequency = float(line.split(':')[1])
+            elif "coordinate system" in line.lower():
+                coordinate_sys = line.split(':')[1].strip()
+            elif "efficiency" in line.lower():
+                antenna_efficiency = float(line.split(':')[1])
+
+        # Print the metadata and the first 10 x, y, and ffp values
+        print('Antenna name:', antenna_name)
+        print('Antenna type:', pattern_type)
+        # print('X:', x[:10])
+        # print('Y:', y[:10])
+        # print('FFP:', ffp[:10])
+        if coordinate_sys == "Spherical":
+            datafile = np.loadtxt(file_path, comments=['#', '*'])
+            el = 90 - datafile[:, 0]
+            az = datafile[:, 1]
+            # x,y,z=MF.polar2cart(theta, phi, np.ones(theta.shape))
+            # azimuth,elevation,r=MF.cart2sph(x,y,z)
+            elev_res = np.unique(el).size
+            az_res = np.unique(az).size
+            if pattern_type == "Directivity":
+                etheta_field = np.flipud((datafile[:, 2] + 1j * datafile[:, 3]).reshape(az_res, elev_res).transpose())
+                ephi_field = np.flipud((datafile[:, 4] + 1j * datafile[:, 5]).reshape(az_res, elev_res).transpose())
+
+            # elevation=np.flipud(GF.thetatoelevation(theta).reshape(phi_res, theta_res).transpose())
+            # azimuth=phi.reshape(phi_res, theta_res).transpose()
+            # etheta=etheta_field
+            # ephi=np.flipud(ephi_field)
+        self.pattern_frequency=frequency
+        self.azimuth_resolution=az_res
+        self.elevation_resolution=elev_res
+        self.az_mesh = np.flipud(az.reshape(az_res, elev_res).transpose())
+        self.elev_mesh_mesh = np.flipud(el.reshape(az_res, elev_res).transpose())
+        self.pattern[:, :, 0] = copy.deepcopy(etheta_field)
+        self.pattern[:, :, 1] = copy.deepcopy(ephi_field)
+        self.display_pattern(desired_pattern="Power")
+
+        return True
+
+    def export_ffe(self, path, source_name="Antenna Export"):
+        import lyceanem
+        destination_path = path.with_suffix(".ffe")
+        filetype = "Far Field"
+        fileformat = "8"
+        import datetime
+        # Get the current date and time
+        now = datetime.datetime.now()
+        # Format the date and time as a string
+        date_string = now.strftime("%Y-%m-%d %H:%M:%S")
[…]
