+++ lyceanem/raycasting/rayfunctions.py
@@ -2294,32 +2294,100 @@ def chunkingRaycaster1Dv3(
     sources, sinks, scattering_points, filtered_index, environment_local, terminate_flag
 ):
     start = timer()
-    sink_index = np.arange(
-        sources.shape[0] + 1, sources.shape[0] + 1 + sinks.shape[0]
-    ).reshape(sinks.shape[0], 1)
-    scattering_point_index = np.arange(
-        np.max(sink_index) + 1, np.max(sink_index) + 1 + scattering_points.shape[0]
-    ).reshape(scattering_points.shape[0], 1)
-    if not terminate_flag:
-        target_indexing = create_model_index(
-            filtered_index, sink_index, scattering_point_index
-        )
-    else:
-        target_indexing = create_model_index(
-            filtered_index, sink_index, np.empty((0, 0), dtype=np.int32)
-        )  # only target rays at sinks
-
-    prep_dt = timer() - start
-    raystart = timer()
-    RAYS_CAST = 0
-    # the rays must fit in GPU memory, aim for no more than 80% utilisation
-    # establish memory limits
     free_mem, total_mem = cuda.current_context().get_memory_info()
     max_mem = np.ceil(free_mem * 0.5).astype(np.int64)
     ray_limit = (
         np.floor(np.floor((max_mem - environment_local.nbytes) / base_types.ray_t.size) / 1e7)
         * 1e7
     ).astype(np.int64)
+    sink_index = np.arange(
+        sources.shape[0] + 1, sources.shape[0] + 1 + sinks.shape[0]
+    ).reshape(sinks.shape[0], 1)
+    scattering_point_index = np.arange(
+        np.max(sink_index) + 1, np.max(sink_index) + 1 + scattering_points.shape[0]
+    ).reshape(scattering_points.shape[0], 1)
+    prep_dt = timer() - start
+    raystart = timer()
+    RAYS_CAST = 0
+    if not terminate_flag:
+        target_indexing = create_model_index(
+            filtered_index, sink_index, scattering_point_index
+        )
+    else:
+        if (filtered_index.shape[0]*sink_index.shape[0])<ray_limit:
+            target_indexing = create_model_index(
+                filtered_index, sink_index, np.empty((0, 0), dtype=np.int32)
+            )  # only target rays at sinks
+        else:
+            # the rays must fit in GPU memory, aim for no more than 80% utilisation
+            # establish memory limits
+            sub_filtered_index = np.array_split(
+                filtered_index, np.ceil(filtered_index.shape[0]*sink_index.shape[0] / ray_limit).astype(int)
+            )
+            chunknum = len(sub_filtered_index)
+            filtered_index2 = np.empty((0, filtered_index.shape[1]+1), dtype=np.int32)
+            final_index2 = np.empty((0, filtered_index.shape[1]+1), dtype=np.int32)
+            # print('chunking total of ',target_indexing.shape[0],' rays in ',chunknum,' batches')
+            for chunkindex in range(chunknum):
+                sub_target=create_model_index(
+                sub_filtered_index[chunkindex], sink_index, np.empty((0, 0), dtype=np.int32)
+                )
+                # cycle the raycaster over the sub arrays
+                threads_in_block = 256
+                second_ray_payload = charge_rays_environment1Dv2(
+                    sources, sinks, scattering_points, sub_target
+                )
+                # ray_chunks=np.linspace(0,ray_num,math.ceil(ray_num/chunk_size)+1,dtype=np.int32)
+                # for idx in range(len(triangle_chunk)-1):
+                d_environment = cuda.device_array(
+                    len(environment_local), dtype=base_types.triangle_t
+                )
+                cuda.to_device(environment_local, to=d_environment)
+                d_chunk_payload = cuda.device_array(
[…]
