### 2023-03-29 #5234 @arfon
Hi @sasha-seneca – no problem at all. Thank you so much for all of your work thus far. Are there any other items you'd recommend the author address (or issues you've opened on the upstream repository (https://github.com/LyceanEM/LyceanEM-Python))?

Also, I realise this is a big ask – but do you know anyone else who might be able to assist us with providing a second review? We've been struggling to find people with the right background (and now also with access to the appropriate hardware it would seem).

Thanks in advance!
