### 2023-03-29 #5234 @LyceanEM
Hi @arfon @sasha-seneca

I am happy to say that although the most effective computing architecture is expected to remain CUDA, implementing CPU modes and more cross platform support is a key aspect of the planned development over the next year. I am currently preparing the vision statement for LyceanEM to lay out the roadmap for new features for the user base.

Thank you for your work on this review. If there is anything I can do to assist, please let me know.
