### 2023-04-27 #5234 @LyceanEM
@arfon Have you had any luck finding reviewers? I have not had much luck myself. Although perhaps @craig-warren might be willing to review, as he maintains gprMax (https://github.com/gprMax/gprMax), which has some similarities, although it is a more mature FDTD based code.
