### 2023-04-27 1f0a59b
Bugfix for mesh_conversion and added import and export functions from antenna pattern class to .ffe format for FEKO. Also added JOSS Badge
