### 2023-04-29 #5234 @arfon
> @arfon Have you had any luck finding reviewers? I have not had much luck myself.

I'm afraid I haven't sorry.

> Although perhaps @craig-warren might be willing to review, as he maintains gprMax (https://github.com/gprMax/gprMax), which has some similarities, although it is a more mature FDTD based code.

Definitely could be an option. Let's see if the ping on GitHub here gets his attention, otherwise I will send him an email directly.

@craig-warren – would you be willing to review this submission for JOSS? The submission under consideration is LyceanEM, a A Python package for virtual prototyping of antenna arrays, time and frequency domain channel modelling (https://github.com/LyceanEM/LyceanEM-Python)

Many thanks
Arfon
