### 2023-05-02 #5234 @craig-warren
@arfon, yes I'd be happy to review. I've not done a JOSS review before but I see their are guidelines which I will read over in the next day or so.
