### 2023-05-07 #5234 @arfon
> @arfon, yes I'd be happy to review. I've not done a JOSS review before but I see their are guidelines which I will read over in the next day or so.

Wonderful, thank you so much @craig-warren. I'll go ahead and set you up as a reviewer now.
