+++ docs/source/paper.bib
@@ -17,6 +17,7 @@
   journal  = {IEEE Access},
   year     = {2021},
   volume   = {9},
+  doi      = {10.1109/ACCESS.2021.3074317},
   note     = {Publisher: IEEE},
   abstract = {Conformal antenna arrays have long represented an attractive solution to challenges imposed by an electromagnetic aperture on an aerospace platform. This allows designers to trade between array positioning and geometry, and avoids the significant aerodynamic drag of a planar array. This is hampered by the expense of the design and computational modelling required to ensure that a conformal array w...},
   url      = {https://ieeexplore.ieee.org/document/9408625/},
@@ -26,6 +27,7 @@
   author    = {Timothy Pelham},
   title     = {Rapid Antenna and Array Analysis for Virtual Prototyping},
   booktitle = {{Radar} 2022},
+  doi       = {10.1049/icp.2022.2330},
   year      = {2022},
   abstract  = {The increased use of radar systems in new form factors from automotive to small unmanned aerial vehicles requires the development of robust and accessible methods for virtual prototyping. Allowing a project team to evaluate the influence of platform structure, antenna or array geometry and beamforming architecture on overall system performance. Using spatial measurements of a Sea Searcher X band radar as an example, LyceanEM is used to predict the influence of reflector shape on maximum achievable directivity, and predicted antenna pattern, compared with full electromagnetic solver (CST) at 9.3GHz. LyceanEM showed a correlation of 0.8 between the predicted and full solver antenna pattern, predicting a maximum directivity of 19.36dBi, compared to 21.97dBi for CST, using 109 seconds of runtime, versus 37 hours for the time domain solver.},
 }
