### 2023-06-01 #5234 @craig-warren
@LyceanEM @arfon , I was finally able to start the review process today. A have a few comments (see below) but generally looks good so far. I have to find NVIDIA GPU to test functionality which I hope to do next week.

**State of the field**
As already mentioned by @sasha-seneca, it would be useful to know what other software packages provide similar functionality. This would give the reader a better understanding of where LyceanEM fits in the EM software landscape.

**References**
Full reference needed for ‘Rapid antenna and array analysis for virtual prototyping’ - page numbers and DOI missing.

**Tests**
It would be useful to mention in the documentation (at a high level) the tests you have implemented using pytest
