### 2023-06-01 #5234 @LyceanEM
@craig-warren
**State of the field**
I am not aware of any other program providing similar functionality. The motivation behind creating this software and publishing it was the identification of a need for a method for rapid virtual prototyping which is not covered by commercial solvers like CST, HFSS, FEKO etc, based upon my PhD research. The channel modelling functionality was added later.

**References**
I have updated the paper today with all the missing DOI, when I submitted not all were avaliable. I will add the page numbers as well today.

**Tests**
There is at the moment very limited test coverage via pytest, as I did not start development using the test driven development model. At the moment only the base uvn to xyz excitation functions, and some rotation functions are covered with automated testing. A research software engineer is joining the development team in late June, and as we move to support multiple computing architectures and other features of our development roadmap the intention is to convert fully to a test driven development process, and ideally complete test coverage.
