+++ docs/source/paper.bib
@@ -6,6 +6,7 @@
   publisher = {IEEE},
   abstract  = {The use of millimetre waves (mmWave) requires the development of novel, robust over the air (OTA) test methods. One low-cost OTA method is based upon an elliptical arrangement of reflectors, to allow conformance testing of devices at mmWave. A discrete ray model is used to predict the influence of reflector shape and arrangement on scattering at 24GHz for two reference surfaces of differing size and a substrate reflector. The model shows a correlation of 0.99 between the predicted and measured scattering for the larger reference surface, and root mean square (RMS) error of -69dB, while for the smaller reference surface the model achieved a correlation of 0.98 between the predicted and measured scattering, and an RMS error of -69dB.},
   doi       = {10.23919/EuCAP51087.2021.9410981},
+  pages     = {1-4},
   isbn      = {978-88-312-9902-2},
   keywords  = {antenna, electromagnetic, measurements, propagation},
   url       = {https://ieeexplore.ieee.org/document/9410981/},
@@ -17,6 +18,7 @@
   journal  = {IEEE Access},
   year     = {2021},
   volume   = {9},
+  pages    = {60880-60890},
   doi      = {10.1109/ACCESS.2021.3074317},
   note     = {Publisher: IEEE},
   abstract = {Conformal antenna arrays have long represented an attractive solution to challenges imposed by an electromagnetic aperture on an aerospace platform. This allows designers to trade between array positioning and geometry, and avoids the significant aerodynamic drag of a planar array. This is hampered by the expense of the design and computational modelling required to ensure that a conformal array w...},
@@ -24,12 +26,17 @@
 }
 
 @InProceedings{Pelham:2022,
-  author    = {Timothy Pelham},
-  title     = {Rapid Antenna and Array Analysis for Virtual Prototyping},
-  booktitle = {{Radar} 2022},
-  doi       = {10.1049/icp.2022.2330},
-  year      = {2022},
-  abstract  = {The increased use of radar systems in new form factors from automotive to small unmanned aerial vehicles requires the development of robust and accessible methods for virtual prototyping. Allowing a project team to evaluate the influence of platform structure, antenna or array geometry and beamforming architecture on overall system performance. Using spatial measurements of a Sea Searcher X band radar as an example, LyceanEM is used to predict the influence of reflector shape on maximum achievable directivity, and predicted antenna pattern, compared with full electromagnetic solver (CST) at 9.3GHz. LyceanEM showed a correlation of 0.8 between the predicted and full solver antenna pattern, predicting a maximum directivity of 19.36dBi, compared to 21.97dBi for CST, using 109 seconds of runtime, versus 37 hours for the time domain solver.},
+   author = {T. G. Pelham},
+   keywords = {electromagnetic solver;unmanned aerial vehicles;beamforming architecture;reflector shape;virtual prototyping array analysis;LyceanEM;time domain solver;CST;radar systems;maximum achievable directivity;full solver antenna pattern;Sea Searcher X band radar;},
+   language = {English},
[…]
