+++ README.md
@@ -62,8 +62,24 @@ Assuming the cudatoolkit and cupy are already installed, then LyceanEM can also 
 ```
  pip install LyceanEM
 ```
+## Development Roadmap
+
+`LyceanEM` is electromagnetics simulation software that is used by researchers and engineers in a variety of fields. The software is currently under development, and the developers have outlined a roadmap for future changes. The roadmap includes three key areas:
+
+* Computational efficiency and scalability: The developers plan to improve the computational efficiency of `LyceanEM` so that it can be used on a wider range of hardware platforms, including desktop computers and high-performance computing (HPC) clusters. This will make `LyceanEM` more accessible to a wider range of users. It is the intention of the developers to support antenna arrays with multiple billon antenna elements, `Giga-scale` antenna arrays.
+* Core propagation engine: The developers plan to improve the core propagation engine of `LyceanEM` to include more realistic models of lossy propagation, atmospheric effects for each layer of the atmosphere, and dynamic environments. This will make `LyceanEM` more accurate and versatile for a wider range of applications.
+* Modelling fidelity: The developers plan to add new features to `LyceanEM` that will allow users to model electromagnetic systems with greater fidelity. This includes support for importing antenna patterns and time domain sources, as well as the development of open source standards for antenna array designs, antenna patterns and field sources, and wireless power transfer.
+
+Here are some specific ways that users can contribute to the development of `LyceanEM`:
+
+* Report bugs: If you find a bug in `LyceanEM`, please report it to the developers so that they can fix it.
+* Submit patches: If you know how to fix a bug or add a new feature, please submit a pull request to the developers.
+* Donate: If you would like to support the development of `LyceanEM`, you can make a donation to the developers.
+
+Your contributions will help to make `LyceanEM` the best possible electromagnetics simulation software for a wide range of users. Thank you for your support!
+
 ## Resources
 
 * Code: [github.com/LyceanEM/LyceanEM-Python](https://github.com/LyceanEM/LyceanEM-Python)
-* Documentation: [lyceanem-python.readthedocs.io/en/latest/](https://lyceanem-python.readthedocs.io/en/latest/)
+* Documentation: [https://documentation.lyceanem.com/en/latest/index.html](https://documentation.lyceanem.com/en/latest/index.html)
 * License: [github.com/LyceanEM/LyceanEM-Python/blob/master/LICENSE.txt](https://github.com/LyceanEM/LyceanEM-Python/blob/master/LICENSE.txt)
