+++ docs/source/contributing.rst
@@ -3,7 +3,7 @@ Contributing
 =============
 The `LyceanEM` project welcomes interested contributors. This isn't limited to programming, and additional examples, documention, or demonstrations are welcomed.
 
-If you are unsure where to start or how you can contribute reach out on the `GitHub <https://github.com/`LyceanEM`/`LyceanEM`-Python>`_ by opening an issue or commenting on an issue you are interested in.
+If you are unsure where to start or how you can contribute reach out on the `GitHub <https://github.com/LyceanEM/LyceanEM-Python>`_ by opening an issue or commenting on an issue you are interested in.
 
 All accepted contributions must be accepted under the :ref:`GNU 3 license<copyright>`.
 
