### 2023-06-06 8b6d28d
Added passthrough for aperture projection to set the desired sample points radius from the aperture origin.
