### 2023-06-06 #5234 @craig-warren
@LyceanEM I have successfully installed LyceanEM on my desktop Windows 10 machine which has external NVIDIA (1080 & 2080) GPUs attached. However, I encountered several errors when testing the examples.

Examples 01, 02, 03, 04, 05 and 08 would not fully complete with the following typical traceback:

```
Traceback (most recent call last):
  File "01_aperture_projection.py", line 93, in <module>
    directivity_envelope, pcd = aperture_projection(
  File "C:\Users\cwarren\miniconda3\envs\LyceanEM\lib\site-packages\lyceanem\models\frequency_domain.py", line 49, in aperture_projection
    blocking_triangles=GF.mesh_conversion(environment)
  File "C:\Users\cwarren\miniconda3\envs\LyceanEM\lib\site-packages\lyceanem\geometry\geometryfunctions.py", line 267, in mesh_conversion
    triangles = object.triangles_base_raycaster()
AttributeError: type object 'object' has no attribute 'triangles_base_raycaster'
```

Example 06 had the following traceback:
```
Structure does not exist
C:\Users\cwarren\miniconda3\envs\LyceanEM\lib\site-packages\lyceanem\electromagnetics\beamforming.py:1252: RuntimeWarning: divide by zero encountered in log10
  logdata = 10 * np.log10(data)
Structure does not exist
Structure does not exist
Structure does not exist
Structure does not exist
```
Example 07 had the following traceback:
```
Traceback (most recent call last):
  File "07_aperture_farfield_polarisation.py", line 51, in <module>
    Etheta, Ephi = calculate_farfield(
  File "C:\Users\cwarren\miniconda3\envs\LyceanEM\lib\site-packages\lyceanem\models\frequency_domain.py", line 147, in calculate_farfield
    environment_triangles=GF.mesh_conversion(antenna_solid)
  File "C:\Users\cwarren\miniconda3\envs\LyceanEM\lib\site-packages\lyceanem\geometry\geometryfunctions.py", line 269, in mesh_conversion
    exported_structure=base_classes.structures(solids=object.export_all_structures())
AttributeError: type object 'object' has no attribute 'export_all_structures'
```

I also found I needed to install the following packages across different examples: `tqdm, folium, skyfield, geopandas`
