### 2023-06-06 #5234 @LyceanEM
Hello @craig-warren

Thanks for running the examples, can you try installing lyceanem from source and seeing if this corrects the issue? I think this is due to a bug in the `mesh_conversion' which was reported and fixed. The latest code on github should run all these examples without issues, but I haven't issued a new pypi version since the bugfix.

Example 06 is correctly reporting that there is no structure to block rays provided to the raycaster, which is correct behaviour. It is a somewhat clunky message to the user currently, but it was thought preferable to flag that there are no structures.

Many Thanks

Tim
