### 2023-06-06 #5234 @LyceanEM
Hi @craig-warren

Just as an update, I have tried the examples with the latest source code (which is the version when installing from github via pip), and all the examples run without errors, except for the warning in Example 06.

Kind Regards

Tim
