### 2023-06-06 #5234 @craig-warren
@LyceanEM thanks, yes I managed to run all examples (except 08) successfully from the source code. Example 08 is now giving me the following traceback:

```
Traceback (most recent call last):
  File "08_beamforming_architecture.py", line 104, in <module>
    directivity_map_discrete = MaximumDirectivityMapDiscrete(
  File "C:\Users\cwarren\Downloads\LyceanEM-Python\lyceanem\electromagnetics\beamforming.py", line 627, in MaximumDirectivityMapDiscrete
    WS_weights = WeightTruncation(WS_weights, resolution)
  File "C:\Users\cwarren\miniconda3\envs\LyceanEM\lib\site-packages\numba\core\dispatcher.py", line 468, in _compile_for_args
    error_rewrite(e, 'typing')
  File "C:\Users\cwarren\miniconda3\envs\LyceanEM\lib\site-packages\numba\core\dispatcher.py", line 409, in error_rewrite
    raise e.with_traceback(None)
numba.core.errors.TypingError: Failed in nopython mode pipeline (step: nopython frontend)
No implementation of function Function(<function round_ at 0x000002A201D58160>) found for signature:

 >>> round_(array(complex128, 1d, C), Literal[int](0), array(float32, 1d, C))

There are 2 candidate implementations:
  - Of which 2 did not match due to:
  Overload in function 'impl_np_round': File: numba\np\arraymath.py: Line 3125.
    With argument(s): '(array(complex128, 1d, C), int64, array(float32, 1d, C))':
   Rejected as the implementation raised a specific error:
     TypingError: Failed in nopython mode pipeline (step: nopython frontend)
   No implementation of function Function(<built-in function setitem>) found for signature:

    >>> setitem(array(float32, 1d, C), UniTuple(int64 x 1), complex128)

   There are 16 candidate implementations:
         - Of which 16 did not match due to:
         Overload of function 'setitem': File: <numerous>: Line N/A.
           With argument(s): '(array(float32, 1d, C), UniTuple(int64 x 1), complex128)':
          No match.

   During: typing of setitem at C:\Users\cwarren\miniconda3\envs\LyceanEM\lib\site-packages\numba\np\arraymath.py (3177)

   File "..\..\miniconda3\envs\LyceanEM\lib\site-packages\numba\np\arraymath.py", line 3177:
               def impl(a, decimals=0, out=None):
                   <source elided>
                   for index, val in np.ndenumerate(a):
                       out[index] = np.round(val, decimals)
                       ^

  raised from C:\Users\cwarren\miniconda3\envs\LyceanEM\lib\site-packages\numba\core\typeinfer.py:1086
During: resolving callee type: Function(<function round_ at 0x000002A201D58160>)
During: typing of call at C:\Users\cwarren\Downloads\LyceanEM-Python\lyceanem\electromagnetics\beamforming.py (1125)

File "..\LyceanEM-Python\lyceanem\electromagnetics\beamforming.py", line 1125:
def WeightTruncation(weights, resolution):
    <source elided>
    # shift numpy complex angles from -pi to pi, into 0 to 2pi, then quantise for `quant' levels
    levels = np.round(
```
