### 2023-06-06 #5234 @LyceanEM
Hi @craig-warren

That is more difficult to identify, the example runs fine on my platform. It looks like numba is failing to compile WeightTruncation due to variable typing not matching the candidate implemntations. I can take a look at that this afternoon and issue an update to the source code? I think this has to be a numba version specific error, what version of numba are you using? In my normal environment I have numba version 0.56.4.
