### 2023-06-06 #5234 @craig-warren
Hi @LyceanEM The version of numba was 0.57.0. I had to use Python 3.8.x to get open3d to install, it didn't work with Python 3.11.x.
