### 2023-06-06 #5234 @LyceanEM
> Hi @LyceanEM The version of numba was 0.57.0. I had to use Python 3.8.x to get open3d to install, it didn't work with Python 3.11.x.

Right, I will check with this version now. Open3d is a bit difficult as a dependancy, which is why I am moving to another backend for 3D file format support and object manipulation. I will likely then publish an optional package using open3d for visualisations.
