### 2023-06-06 8aacaff
Added Bugfix to numba type inference in WeightTruncation. This ensures that Numba 0.57 will not raise an error but calculate the truncated weights correctly.
