### 2023-06-06 #5234 @LyceanEM
@craig-warren I have now written and tested a bugfix for this behaviour, and pushed it to the github. If you now install the latest source code it should work without error.
