### 2023-06-07 #5234 @craig-warren
@LyceanEM I pulled the latest source and re-ran example 08. However, it seems to be just continually running on the GPU, i.e. hangs.
