### 2023-06-07 #5234 @LyceanEM
It is an example that takes a while to run as the beamforming comparison is cpu based right now. If you reduce the farfield pattern resolution (az_res, elev_res) it should run much quicker.
