### 2023-06-07 #5234 @LyceanEM
Also it's worth noting that the open3d windows are blocking so nothing happens until the window is closed.
