### 2023-06-07 #5234 @craig-warren
@LyceanEM - yep got it now, reduced the farfield resolution as you suggested.

@arfon @LyceanEM I am happy and have nothing further to add to review.
