### 2023-06-12 #5234 @LyceanEM
@arfon Thank you, I can confirm I have issued a new release, version 0.0.6, and archived in Zenodo with DOI 10.5281/zenodo.8026567.
