### 2023-06-12 029e2c2
Bugfix for antenna_structure rotation function, causing subobjects to be rotated twice.
