### 2023-06-12 #5234 @kyleniemeyer
Hi @LyceanEM, a few things before publishing:
- [ ] remove the "These authors contributed equally." second footnote in the author section
- [ ] could you add that missing DOI? It does look correct for the conference paper.
- [ ] while you are in the references file, can you use first initial rather than your full name for the final reference? (meaning, "T. G. Pelham" rather than "Timothy G. Pelham")
- [ ] some minor typos: "This setup is show in Figure 5" on page 3, and an extraneous comma before "shows" in the paragraph explaining Figure 6 on page 4.
