+++ docs/source/paper.bib
@@ -13,7 +13,7 @@
 }
 
 @Article{Pelhamb:2021,
-  author   = {Pelham, Timothy G. and Hilton, Geoff and Mellios, Evangelos and Lewis, Rob},
+  author   = {Pelham, T. G. and Hilton, Geoff and Mellios, Evangelos and Lewis, Rob},
   title    = {Conformal {Antenna} {Array} {Design} {Using} {Aperture} {Synthesis} and {On}-{Platform} {Modeling}},
   journal  = {IEEE Access},
   year     = {2021},
@@ -26,7 +26,7 @@
 }
 
 @InProceedings{Pelham:2022,
-   author = {T. G. Pelham},
+   author = {Pelham, T.G.},
    keywords = {electromagnetic solver;unmanned aerial vehicles;beamforming architecture;reflector shape;virtual prototyping array analysis;LyceanEM;time domain solver;CST;radar systems;maximum achievable directivity;full solver antenna pattern;Sea Searcher X band radar;},
    language = {English},
    abstract = {The increased use of radar systems in new form factors from automotive to small unmanned aerial vehicles requires the development of robust and accessible methods for virtual prototyping. Allowing a project team to evaluate the influence of platform structure, antenna or array geometry and beamforming architecture on overall system performance. Using spatial measurements of a Sea Searcher X band radar as an example, LyceanEM is used to predict the influence of reflector shape on maximum achievable directivity, and predicted antenna pattern, compared with full electromagnetic solver (CST) at 9.3GHz. LyceanEM showed a correlation of 0.8 between the predicted and full solver antenna pattern, predicting a maximum directivity of 19.36dBi, compared to 21.97dBi for CST, using 109 seconds of runtime, versus 37 hours for the time domain solver.},
@@ -34,6 +34,7 @@
    journal = {IET Conference Proceedings},
    year = {2022},
    month = {10},
+   doi = {10.1049/icp.2022.2330},
    pages = {278-282(4)},
    publisher ={Institution of Engineering and Technology},
    url = {https://digital-library.theiet.org/content/conferences/10.1049/icp.2022.2330},
