### 2023-06-12 8c3788b
Adding Citation file to github to improve ease of citation.
