### 2023-06-12 7433b18
Adding Citation file to github to improve ease of citation.
