+++ README.md
@@ -44,6 +44,9 @@ The package has been tested on:
 With Python versions:
 
 * 3.7
+* 3.8
+* 3.9
+* 3.10
 
 ## Installation
 
